separating entry and entry list and adding name-value pair conversion
